# docs/openapi.yaml
openapi: 3.1.0
info:
  title: "Budget Free Engine API"
  description: |
    Contrato de API gerado pelo SystemForge a partir do LLD.md.
    Sistema de geração de orçamentos para projetos de software — lead magnet.
    Monolito Next.js App Router com Prisma + Supabase PostgreSQL.
  version: "1.0.0"
  contact:
    name: "Budget Free Engine Team"
servers:
  - url: "http://localhost:3000/api/v1"
    description: "Desenvolvimento"
  - url: "https://orcamentogratuito.tech/api/v1"
    description: "Produção"
security:
  - supabaseAuth: []
...
components:
  securitySchemes:
    supabaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Token JWT emitido pelo Supabase Auth (magic link). Usado apenas em rotas /admin/*."
  schemas:
...
    LeadSummary:
      type: object
      description: "Resumo do lead para exibição no admin panel"
      required:
        - id
        - name
        - email
        - project_type
        - score
        - complexity
        - estimated_price_min
        - estimated_price_max
        - estimated_days_min
        - estimated_days_max
        - created_at
      properties:
        id:
          type: string
          example: "lead_abc123"
        name:
          type: string
          example: "João Silva"
        email:
          type: string
          format: email
          example: "joao@example.com"
        project_type:
          $ref: '#/components/schemas/ProjectType'
        score:
          $ref: '#/components/schemas/LeadScore'
        complexity:
          $ref: '#/components/schemas/ComplexityLevel'
        estimated_price_min:
          type: number
          format: float
          example: 15000.0
        estimated_price_max:
          type: number
          format: float
          example: 25000.0
        estimated_days_min:
          type: integer
          example: 45
        estimated_days_max:
          type: integer
          example: 60
        created_at:
          type: string
          format: date-time
          example: "2026-03-21T10:00:00Z"

    # ──────────────────────────────────────────────────────────────────────────
    # Schemas de Input (Request Bodies)
    # ──────────────────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: "Código de erro único"
              example: "SESSION_NOT_FOUND"
            message:
              type: string
              description: "Mensagem legível do erro"
              example: "Sessão não encontrada."
            details:
              type: [string, 'null']
              description: "Informações adicionais de validação (opcional)"
              example: null
    RateLimitErrorResponse:
      type: object
      required:
        - error
        - retry_after
      properties:
        error:
          type: string
          example: "Too many requests. Try again later."
        retry_after:
          type: integer
          description: "Segundos até poder tentar novamente"
          example: 60

    # ──────────────────────────────────────────────────────────────────────────
    # Enums
    # ──────────────────────────────────────────────────────────────────────────
...
    ProjectType:
      type: string
      enum:
        - web_system
        - landing_page
        - mobile_app
        - saas_mvp
        - e_commerce
        - automation
      description: "Tipo de projeto selecionado pelo visitante em Q001"
      example: "web_system"
...
    LeadScore:
      type: string
      enum: [A, B, C]
      description: |
        Score de qualificação do lead:
        - A: score_total >= 70 (alta prioridade)
        - B: score_total >= 40 (média prioridade)
        - C: score_total < 40 (baixa prioridade)
      example: "A"
    ComplexityLevel:
      type: string
      enum: [low, medium, high, very_high]
      description: |
        Nível de complexidade inferido do complexity_score acumulado:
        - low: 0-30
        - medium: 31-50
        - high: 51-70
        - very_high: 71+
      example: "medium"
...
  responses:
    UnauthorizedError:
      description: "Não autenticado — token ausente, inválido ou expirado (401)"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Autenticação necessária."
...
    RateLimitError:
      description: "Rate limit atingido (429)"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitErrorResponse'
          example:
            error: "Too many requests. Try again later."
            retry_after: 60
    InternalError:
      description: "Erro interno do servidor (500)"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "Erro interno. Tente novamente em alguns instantes."

# ──────────────────────────────────────────────────────────────────────────────
# Paths
# ──────────────────────────────────────────────────────────────────────────────
paths:
...
  /admin/leads:
    get:
      tags: [admin]
      summary: "Listar leads (admin)"
      description: |
        Retorna lista paginada de leads com filtros opcionais.
        Exibe dados de contato, score, estimativa e complexidade.
        Requer autenticação via Supabase Auth (magic link).

        **Rate limit:** 10 req/60s por IP.
      security:
        - supabaseAuth: []
      parameters:
        - name: score
          in: query
          required: false
          description: "Filtrar por score de qualificação"
          schema:
            $ref: '#/components/schemas/LeadScore'
          example: "A"
        - name: type
          in: query
          required: false
          description: "Filtrar por tipo de projeto (ProjectType)"
          schema:
            type: string
          example: "web_system"
        - name: from
          in: query
          required: false
          description: "Data início do filtro (ISO 8601)"
          schema:
            type: string
            format: date-time
          example: "2026-01-01T00:00:00Z"
        - name: to
          in: query
          required: false
          description: "Data fim do filtro (ISO 8601)"
          schema:
            type: string
            format: date-time
          example: "2026-03-31T23:59:59Z"
        - name: page
          in: query
          required: false
          description: "Número da página (padrão: 1)"
          schema:
            type: integer
            default: 1
            minimum: 1
          example: 1
        - name: per_page
          in: query
          required: false
          description: "Registros por página (padrão: 20, máximo: 100)"
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          example: 20
      responses:
        '200':
          description: "Lista de leads retornada com sucesso"
          content:
            application/json:
              schema:
                type: object
                required: [data, total, page, per_page]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeadSummary'
                  total:
                    type: integer
                    description: "Total de leads encontrados com os filtros aplicados"
                    example: 42
                  page:
                    type: integer
                    example: 1
                  per_page:
                    type: integer
                    example: 20
              example:
                data:
                  - id: "lead_abc"
                    name: "João Silva"
                    email: "joao@example.com"
                    project_type: "web_system"
                    score: "A"
                    complexity: "medium"
                    estimated_price_min: 15000.0
                    estimated_price_max: 25000.0
                    estimated_days_min: 45
                    estimated_days_max: 60
                    created_at: "2026-03-21T10:00:00Z"
                total: 42
                page: 1
                per_page: 20
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          description: "Parâmetros de filtro inválidos"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "VALIDATION_FAILED"
                  message: "Parâmetros de consulta inválidos."
                  details: "score: Invalid enum value"
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Server Actions (expostos como endpoints internos) ─────────────────────
...
  /cron/anonymize-leads:
    get:
      tags: [admin]
      summary: "Cron: anonimizar leads (LGPD/GDPR — retenção 12 meses)"
      x-internal: true
      security: []
      parameters:
        - { name: Authorization, in: header, required: true, schema: { type: string } }
      responses:
        '200':
          description: "Anonimização concluída em batches de 100"
          content:
            application/json:
              schema:
                type: object
                properties:
                  anonymized: { type: integer }
                  duration_ms: { type: integer }
        '401':
          description: "CRON_SECRET inválido"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /cron/send-resume-emails:
    get:
      tags: [admin]
      summary: "Cron: enviar emails de retomada para sessões incompletas"
      x-internal: true
      security: []
      parameters:
        - { name: Authorization, in: header, required: true, schema: { type: string } }
      responses:
        '200':
          description: "Emails enviados"
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: integer }
                  failed: { type: integer }
                  duration_ms: { type: integer }
        '401':
          description: "CRON_SECRET inválido"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'